Refactored all code
