+++ methods/stan_2d_main.py
@@ -5,7 +5,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 from tqdm import tqdm
 from math import floor, ceil
-from methods.stan.stan_2d import stan_2d_model
+from methods.stan_2d import stan_2d_model
 from sklearn.model_selection import train_test_split
 from sklearn.metrics import accuracy_score, confusion_matrix, roc_auc_score, f1_score, average_precision_score
